chore: Move deduplication to service tracker [DHIS-17744]
