+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/deduplication/DeduplicationService.java
@@ -25,26 +25,29 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.deduplication;
+package org.hisp.dhis.tracker.deduplication;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import java.util.ArrayList;
 import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
-@Data
-@Builder
-@JsonDeserialize(builder = MergeObject.MergeObjectBuilder.class)
-@NoArgsConstructor
-@AllArgsConstructor
-public class MergeObject {
-  @Builder.Default @JsonProperty private List<String> trackedEntityAttributes = new ArrayList<>();
+public interface DeduplicationService {
+  PotentialDuplicate getPotentialDuplicateById(long id);
 
-  @Builder.Default @JsonProperty private List<String> relationships = new ArrayList<>();
+  PotentialDuplicate getPotentialDuplicateByUid(String uid);
 
-  @Builder.Default @JsonProperty private List<String> enrollments = new ArrayList<>();
+  int countPotentialDuplicates(PotentialDuplicateCriteria criteria);
+
+  boolean exists(PotentialDuplicate potentialDuplicate) throws PotentialDuplicateConflictException;
+
+  List<PotentialDuplicate> getPotentialDuplicates(PotentialDuplicateCriteria criteria);
+
+  void addPotentialDuplicate(PotentialDuplicate potentialDuplicate)
+      throws PotentialDuplicateConflictException;
+
+  void updatePotentialDuplicate(PotentialDuplicate potentialDuplicate);
+
+  void autoMerge(DeduplicationMergeParams deduplicationRequest)
+      throws PotentialDuplicateConflictException, PotentialDuplicateForbiddenException;
+
+  void manualMerge(DeduplicationMergeParams deduplicationRequest)
+      throws PotentialDuplicateConflictException, PotentialDuplicateForbiddenException;
 }
